Moving flux_stt.py to deepgram/flux/sagemaker/stt.py
